Corrected MVP architecture
