+++ app/src/main/AndroidManifest.xml
@@ -8,7 +8,7 @@
         android:label="@string/app_name_proj2"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".SplashActivity">
+        <activity android:name=".activity.SplashActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
@@ -16,25 +16,13 @@
             </intent-filter>
         </activity>
 
-        <activity android:name=".MainActivity"
+        <activity android:name=".activity.MainActivity"
             android:label="@string/app_name_proj1"
             android:theme="@style/AppTheme">
         </activity>
 
         <activity
-            android:name=".MoviesMainActivity"
-            android:label="@string/app_name_proj2"
-            android:theme="@style/AppTheme"
-            android:configChanges="keyboardHidden|orientation|screenSize"/>
-        <activity
-            android:name=".activity.PopularMoviesMainActivityNew"
-            android:label="@string/app_name_proj2"
-            android:theme="@style/AppTheme"
-            android:configChanges="keyboardHidden|orientation|screenSize"/>
-
-
-        <activity
-            android:name=".MovieDetailActivity"
+            android:name=".activity.MoviesMainActivity"
             android:label="@string/app_name_proj2"
             android:theme="@style/AppTheme"
             android:configChanges="keyboardHidden|orientation|screenSize"/>
