+++ app/src/main/java/com/udacity/myappportfolio/activity/BaseActivity.java
@@ -1,4 +1,4 @@
-package com.udacity.myappportfolio;
+package com.udacity.myappportfolio.activity;
 
 import android.support.v7.app.AppCompatActivity;
