+++ app/src/main/java/com/udacity/myappportfolio/fragment/MovieListFragment.java
@@ -1,10 +1,7 @@
 package com.udacity.myappportfolio.fragment;
 
-import android.content.ContentResolver;
 import android.content.Context;
 import android.content.res.Configuration;
-import android.database.Cursor;
-import android.net.Uri;
 import android.os.Bundle;
 import android.support.annotation.Nullable;
 import android.support.v7.widget.GridLayoutManager;
@@ -24,15 +21,14 @@ import android.widget.TextView;
 import com.rey.material.widget.ProgressView;
 import com.udacity.myappportfolio.R;
 import com.udacity.myappportfolio.adapter.MovieRecyclerViewAdapter;
-import com.udacity.myappportfolio.db.FavMovieContract;
 import com.udacity.myappportfolio.model.Movie;
 import com.udacity.myappportfolio.model.MovieMainBean;
-import com.udacity.myappportfolio.net.MovieDBResponseListener;
-import com.udacity.myappportfolio.net.MovieRestAPI;
-import com.udacity.myappportfolio.net.TheMovieDBClient;
+import com.udacity.myappportfolio.presenter.PopularMoviePresenter;
+import com.udacity.myappportfolio.presenter.PopularMoviePresenterImpl;
 import com.udacity.myappportfolio.util.Constants;
 import com.udacity.myappportfolio.util.MyUtil;
 import com.udacity.myappportfolio.util.RecyclerItemClickListener;
+import com.udacity.myappportfolio.view.MovieMainView;
 
 import java.io.IOException;
 import java.util.ArrayList;
@@ -46,7 +42,7 @@ import jp.wasabeef.recyclerview.animators.adapters.AlphaInAnimationAdapter;
 import jp.wasabeef.recyclerview.animators.adapters.ScaleInAnimationAdapter;
 
 
-public class MovieListFragment extends BaseFragment implements MovieDBResponseListener {
+public class MovieListFragment extends BaseFragment implements MovieMainView {
 
     @Bind(R.id.list)
     RecyclerView lv_gridList;
@@ -75,16 +71,40 @@ public class MovieListFragment extends BaseFragment implements MovieDBResponseLi
     private boolean loading = true;
     private int pastVisiblesItems, visibleItemCount, totalItemCount;
     private GridLayoutManager mLayoutManager;
-    private MovieRestAPI myService;
-    private String sort_by = Constants.POPULARITY;
-    private int page = 1;
+    private String sortBy = Constants.POPULARITY;
+    private int pageNo = 1;
     private ArrayList<Movie> dynamicResultList = new ArrayList<>();
     private MovieRecyclerViewAdapter newsListViewAdapter;
-    private TheMovieDBClient client;
     private OnItemSelectedListener itemSelectlistener;
+    private PopularMoviePresenter presenter;
 
     public interface OnItemSelectedListener {
         void itemSelected(Movie movie);
+    }
+
+    @Override
+    public List<Movie> getMovieList() {
+        return dynamicResultList;
+    }
+
+    public int getPageNo() {
+        return pageNo;
+    }
+
+    public void setPageNo(int pageNo){
+        this.pageNo = pageNo;
+    }
+
+    public String getSortBy(){
+        return sortBy;
+    }
+
+    public void setSortBy(String sortBy){
+        this.sortBy = sortBy;
+    }
+
+    public void isLoading(boolean isloading) {
+        loading = isloading;
     }
 
     public static MovieListFragment getInstance() {
@@ -97,6 +117,8 @@ public class MovieListFragment extends BaseFragment implements MovieDBResponseLi
     public void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setHasOptionsMenu(true);
+
+        presenter = new PopularMoviePresenterImpl(this);
     }
 
     @Override
@@ -117,7 +139,7 @@ public class MovieListFragment extends BaseFragment implements MovieDBResponseLi
             displayMovieList(dynamicResultList);
         }else{
 
-            loadMovieList();
+            presenter.fetchMovieList(getPageNo() , getSortBy());
 
         }
     }
@@ -132,11 +154,10 @@ public class MovieListFragment extends BaseFragment implements MovieDBResponseLi
     @SuppressWarnings("unused")
     @OnClick(R.id.rl_error)
     public void fetchMovieAgain() {
-        dynamicResultList.clear();
-        page = 1;
-        loadMovieList();
+        setPageNo(1);
+        clearMovieList();
+        presenter.fetchMovieList(getPageNo() , getSortBy());
     }
-
 
     private void setUpListeners() {
 
@@ -162,7 +183,7 @@ public class MovieListFragment extends BaseFragment implements MovieDBResponseLi
                         if ((visibleItemCount + pastVisiblesItems) >= totalItemCount) {
                             loading = false;
                             //Do pagination.. i.e. fetch new data
-                            loadMovieList();
+                            presenter.fetchMovieList(getPageNo() , getSortBy());
                         }
                     }
                 }
@@ -193,27 +214,13 @@ public class MovieListFragment extends BaseFragment implements MovieDBResponseLi
     }
 
 
-    private void loadMovieList() {
-        if (dynamicResultList.size() > 0) {
-            showListProgress();
-        } else {
-            showProgress();
-        }
-
-        client = TheMovieDBClient.getInstance();
-
-        client.loadMovies(page,sort_by , this);
-
-    }
-
-
     private void displayMovieList(List<Movie> movieBeanList) {
         if(movieBeanList.size()==0){
             tv_NoFavMovieAddedText.setVisibility(View.VISIBLE);
             allViewGone();
         }else{
             tv_NoFavMovieAddedText.setVisibility(View.GONE);
-            if (page == 2 || setSortByFavChecked) {
+            if (getPageNo() == 2 || setSortByFavChecked) {
                 newsListViewAdapter = new MovieRecyclerViewAdapter(
                         getActivity(), movieBeanList);
                 lv_gridList.setItemAnimator(new FadeInAnimator());
@@ -230,7 +237,17 @@ public class MovieListFragment extends BaseFragment implements MovieDBResponseLi
 
     }
 
-    private void showProgress() {
+    public void clearMovieList() {
+        dynamicResultList.clear();
+    }
+
+    public void addMovieList(ArrayList<Movie> movieList) {
+        dynamicResultList.addAll(movieList);
+    }
+
+
+    @Override
+    public void showProgress() {
         rl_progress.setVisibility(View.VISIBLE);
         rl_gridView.setVisibility(View.GONE);
         rl_error.setVisibility(View.GONE);
@@ -238,7 +255,8 @@ public class MovieListFragment extends BaseFragment implements MovieDBResponseLi
         tv_NoFavMovieAddedText.setVisibility(View.GONE);
     }
 
-    private void showListProgress() {
+    @Override
+    public void showListProgress() {
         rl_progress.setVisibility(View.GONE);
         rl_gridView.setVisibility(View.VISIBLE);
         rl_error.setVisibility(View.GONE);
@@ -246,7 +264,8 @@ public class MovieListFragment extends BaseFragment implements MovieDBResponseLi
         tv_NoFavMovieAddedText.setVisibility(View.GONE);
     }
 
-    private void showGrid() {
+    @Override
+    public void showGrid() {
         rl_progress.setVisibility(View.GONE);
         rl_gridView.setVisibility(View.VISIBLE);
         rl_error.setVisibility(View.GONE);
@@ -254,7 +273,8 @@ public class MovieListFragment extends BaseFragment implements MovieDBResponseLi
         tv_NoFavMovieAddedText.setVisibility(View.GONE);
     }
 
-    private void allViewGone() {
+    @Override
+    public void allViewGone() {
         rl_progress.setVisibility(View.GONE);
         rl_gridView.setVisibility(View.GONE);
         rl_error.setVisibility(View.GONE);
@@ -262,7 +282,8 @@ public class MovieListFragment extends BaseFragment implements MovieDBResponseLi
     }
 
 
-    private void showError(String message) {
+    @Override
+    public void showError(String message) {
 
         loading = true;
 
@@ -330,13 +351,14 @@ public class MovieListFragment extends BaseFragment implements MovieDBResponseLi
 
     @Override
     public void onSuccess(MovieMainBean bean) {
-
         if (bean != null) {
 
             if (bean.getResults() != null && bean.getResults().size() > 0) {
+                int page = getPageNo();
                 page++;
-                loading = true;
-                dynamicResultList.addAll(bean.getResults());
+                setPageNo(page);
+                isLoading(true);
+                getMovieList().addAll(bean.getResults());
                 displayMovieList(bean.getResults());
             } else {
                 showError(getResources().getString(R.string.no_results));
@@ -348,7 +370,6 @@ public class MovieListFragment extends BaseFragment implements MovieDBResponseLi
 
     @Override
     public void onFailure(Throwable t) {
-
         if(t instanceof IOException)
             showError(getResources().getString(R.string.no_internet));
         else
@@ -369,68 +390,36 @@ public class MovieListFragment extends BaseFragment implements MovieDBResponseLi
                 setSortByPopularityChecked = true;
                 setSortByVotingChecked = false;
                 setSortByFavChecked = false;
-                sort_by = Constants.POPULARITY;
-                page = 1;
-                dynamicResultList.clear();
-                loadMovieList();
+                setSortBy(Constants.POPULARITY);
+                setPageNo(1);
+                clearMovieList();
+                presenter.fetchMovieList(getPageNo() , getSortBy());
                 return true;
             }
             if (item.getItemId() == R.id.sort_by_voting) {
                 setSortByVotingChecked = true;
                 setSortByPopularityChecked = false;
                 setSortByFavChecked = false;
-                sort_by = Constants.VOTE_AVERAGE;
-                page = 1;
-                dynamicResultList.clear();
-                loadMovieList();
+                setSortBy(Constants.VOTE_AVERAGE);
+                setPageNo(1);
+                clearMovieList();
+                presenter.fetchMovieList(getPageNo() , getSortBy());
                 return true;
             }
             if (item.getItemId() == R.id.favorites) {
                 setSortByFavChecked = true;
                 setSortByPopularityChecked = false;
                 setSortByVotingChecked = false;
-                ArrayList<Movie> favMovieList = fetchFavMoviesFromDb();
-                dynamicResultList.clear();
-                dynamicResultList.addAll(favMovieList);
+                ArrayList<Movie> favMovieList = presenter.fetchMovieFromDb(getActivity());
+                clearMovieList();
+                addMovieList(favMovieList);
                 displayMovieList(dynamicResultList);
                 return true;
             }
             return false;
         }
     }
-    private ArrayList<Movie> fetchFavMoviesFromDb(){
 
-        ArrayList<Movie> favMovieList = new ArrayList<>();
-
-        Uri uri = FavMovieContract.CONTENT_URI;
-        ContentResolver resolver = getActivity().getContentResolver();
-        String[] projection = new String[]{FavMovieContract.Columns.MOVIE_ID,
-                FavMovieContract.Columns.MOVIE_TITLE,
-                FavMovieContract.Columns.MOVIE_RELEASE_DATE,
-                FavMovieContract.Columns.MOVIE_RATING,
-                FavMovieContract.Columns.MOVIE_SYNOPSIS,
-                FavMovieContract.Columns.MOVIE_POSTER_URL};
-        Cursor cursor =
-                resolver.query(uri, projection,null, null,null);
-
-        if (cursor.moveToFirst()) {
-            do {
-                Movie m = new Movie();
-
-                m.setId(cursor.getInt(0));
-                m.setOriginal_title(cursor.getString(1));
-                m.setRelease_date(cursor.getString(2));
-                m.setVote_average(Double.parseDouble(cursor.getString(3)));
-                m.setOverview(cursor.getString(4));
-                m.setPoster_path(cursor.getString(5));
-
-                favMovieList.add(m);
-
-            } while (cursor.moveToNext());
-        }
-
-        return favMovieList;
-    }
     @Override
     public void onAttach(Context context) {
         super.onAttach(context);
